pyproject.toml
[build-system]
requires = ["flit_core >=3.11,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "nzlusdb"
authors = [
  { name="Baptiste Hamon", email="baptiste.hamon@pg.canterbury.ac.nz" },
]
maintainers = [
  { name="Baptiste Hamon", email="baptiste.hamon@pg.canterbury.ac.nz" },
]
readme = "README.md"
keywords = ["NZLUSDB", "Land Use Suitability", "New Zeland", "Land Suitability Analysis", "Database"]
license = {file = "LICENSE"}
dynamic = ["description", "version"]
dependencies = [
  "h5netcdf >=1.6.0",
  "lsapy >=0.2.0",
  "matplotlib >=3.9.0",
  "netCDF4 >=1.7.0",
  "rioxarray >=0.17.0",
  "xclim >=0.55.1"
]

[project.optional-dependencies]
dev = [
  # dev tools
  "bump-my-version >=1.2.0",
  "codespell >=2.4.1",
  "flit >=3.11",
  "pre-commit >=4.2.0",
  "ruff >=0.12.0",
]
docs = [
  # docs & examples
]
all = ["nzlsudb[dev]", "nzlusdb[docs]"]

[project.urls]
# Homepage = ""
Source = "https://github.com/baptistehamon/nzlusdb"
Issues = "https://github.com/baptistehamon/nzlusdb/issues"
# Zenodo = ""

[tool.bumpversion]
current_version = "1.0-dev1"
commit = true
commit_args = "--no-verify --signoff"
tag = false
allow_dirty = false
parse = "(?P<major>\\d+)\\.(?P<patch>\\d+)(\\-(?P<release>[a-z]+)((?P<build>\\d+)))?"
serialize = [
  "{major}.{patch}-{release}{build}",
  "{major}.{patch}"
]

[[tool.bumpversion.files]]
filename = "src/nzlusdb/__init__.py"
search = "__version__ = \"{current_version}\""
replace = "__version__ = \"{new_version}\""

[tool.bumpversion.parts.build]
independent = false

[tool.bumpversion.parts.release]
optional_value = "release"
values = [
  "dev",
  "release"
]

[tool.ruff]
line-length = 120
output-format = "grouped"

[tool.ruff.format]
docstring-code-format = true
line-ending = "lf"

[tool.ruff.lint]
preview=true
select = [ # default values
  "B", # bugbear
  "C90", # McCabe complexity
  "D", # docstring
  "E", # pycodestyle errors
  "F", # pyflakes
  "I", # imports
  "N", # naming
  "PL", # pylint
  "S", # bandit
  "UP", # pyupgrade
  "W", # pycodestyle warnings
]
ignore = [
  "D400", # end-in-period
  "D401", # non-imperative-mood
  "PLR6201", # literal-membership
]

[tool.ruff.lint.per-file-ignores]
"src/nzlusdb/core/climdataset.py" = ["N803", "N816"]
"src/nzlusdb/suitability/indicators/*.py" = ["N806"]
"src/nzlusdb/core/indicators.py" = ["PLC2701"]
"src/nzlusdb/utils.py" = ["PLR2004"]
# "docs/*.py" = []
# "src/nzlusdb/**/__init__.py" = []
# "tests/**/test*.py" = []

[tool.ruff.lint.isort]
case-sensitive = true
known-first-party = ["nzlusdb"]

[tool.ruff.lint.mccabe]
max-complexity = 20

[tool.ruff.lint.pycodestyle]
max-line-length = 120

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.pylint]
max-args = 15
max-branches = 20
